storage_scanner/requirements.txt
Django
django-extensions
Pillow>=10.2.0
google-cloud-vision>=3.5.0
gunicorn>=21.2.0
python-dotenv>=1.0.0 